# group-7/zhangjunhu/code/backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "M1-QA 投研问答助手 API",
    "description": "对齐 spec/09 API 接口规格 — 9 个端点，支持 SSE 流式问答、会话管理、历史搜索、相似性检测、股票信息查询",
    "version": "1.0.0"
  },
  "servers": [{ "url": "http://localhost:5000", "description": "本地开发" }],
  "tags": [
    { "name": "能力探测", "description": "系统能力状态" },
    { "name": "问答", "description": "问答提交（SSE 流式 + 非流式）" },
    { "name": "会话管理", "description": "会话 CRUD" },
    { "name": "问答记录", "description": "记录查询" },
    { "name": "搜索", "description": "历史搜索与相似性检测" },
    { "name": "股票", "description": "股票信息查询" }
  ],
  "paths": {
    "/api/v1/agent/capabilities": {
      "get": {
        "tags": ["能力探测"],
        "summary": "能力探测",
        "description": "返回当前系统的 LLM 配置状态（CoPaw / 百炼）",
        "operationId": "getCapabilities",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "traceId": { "type": "string", "example": "tr_abc123" },
                    "copaw_configured": { "type": "boolean", "example": false },
                    "bailian_configured": { "type": "boolean", "example": true }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent/ask": {
      "post": {
        "tags": ["问答"],
        "summary": "问答提交（SSE 流式 + 非流式兼容）",
        "description": "当 Accept 头包含 text/event-stream 时返回 SSE 流式响应，否则返回 JSON。三级降级：CoPaw → 百炼 → Demo",
        "operationId": "ask",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["query", "session_id"],
                "properties": {
                  "query": { "type": "string", "minLength": 1, "maxLength": 500, "example": "贵州茅台最新评级是什么" },
                  "session_id": { "type": "string", "format": "uuid", "example": "a414b855-75f9-4108-b534-190e76b62c3d" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "非流式 JSON 响应 / SSE 流式响应",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "traceId": { "type": "string" },
                    "answer": { "type": "string", "example": "贵州茅台最新评级为买入..." },
                    "llm_used": { "type": "boolean", "example": true },
                    "model": { "type": "string", "example": "qwen-plus", "nullable": true },
                    "response_time_ms": { "type": "integer", "example": 1200 },
                    "answer_source": { "type": "string", "enum": ["copaw", "bailian", "demo"] },
                    "stocks": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": { "type": "string", "example": "贵州茅台" },
                          "code": { "type": "string", "example": "600519" }
                        }
                      }
                    }
                  }
                }
              },
              "text/event-stream": {
                "schema": { "type": "string", "description": "SSE 事件流：chunk（部分文本）→ done（完整结果）" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error400" }
        }
      }
    },
    "/api/v1/agent/sessions": {
      "get": {
        "tags": ["会话管理"],
        "summary": "会话列表",
        "description": "返回全部会话，按创建时间倒序",
        "operationId": "getSessions",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "traceId": { "type": "string" },
                    "sessions": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Session" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["会话管理"],
        "summary": "新建会话",
        "operationId": "createSession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string", "default": "新会话", "example": "研报分析会话" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Session" }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent/sessions/{id}": {
      "delete": {
        "tags": ["会话管理"],
        "summary": "删除会话",
        "description": "删除会话并级联删除关联问答记录",
        "operationId": "deleteSession",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "会话 ID" }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "traceId": { "type": "string" },
                    "message": { "type": "string", "example": "删除成功" },
                    "deleted_records": { "type": "integer", "example": 3 }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error400" }
        }
      }
    },
    "/api/v1/agent/sessions/{id}/records": {
      "get": {
        "tags": ["问答记录"],
        "summary": "问答记录列表",
        "description": "返回指定会话的问答记录，按时间正序",
        "operationId": "getRecords",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "会话 ID" }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "traceId": { "type": "string" },
                    "records": { "type": "array", "items": { "$ref": "#/components/schemas/QARecord" } }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error400" }
        }
      }
    },
    "/api/v1/agent/records/search": {
      "get": {
        "tags": ["搜索"],
        "summary": "历史记录搜索",
        "description": "按关键词搜索历史问答记录（匹配 query 和 answer）",
        "operationId": "searchRecords",
        "parameters": [
          { "name": "keyword", "in": "query", "required": true, "schema": { "type": "string" }, "description": "搜索关键词", "example": "茅台" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 }, "description": "返回条数" }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "traceId": { "type": "string" },
                    "records": { "type": "array", "items": { "$ref": "#/components/schemas/QARecord" } }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error400" }
        }
      }
    },
    "/api/v1/agent/records/similar": {
      "post": {
        "tags": ["搜索"],
        "summary": "相似提问检测",
        "description": "检测是否存在历史相似提问（对齐 US-003 AC-003-02）",
        "operationId": "findSimilar",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["query"],
                "properties": {
                  "query": { "type": "string", "example": "贵州茅台的评级是什么" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "traceId": { "type": "string" },
                    "has_similar": { "type": "boolean" },
                    "similar_records": { "type": "array", "items": { "$ref": "#/components/schemas/QARecord" } }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error400" }
        }
      }
    },
    "/api/v1/agent/stock/{code}": {
      "get": {
        "tags": ["股票"],
        "summary": "股票信息查询",
        "description": "查询股票详情（对齐 US-004）",
        "operationId": "getStockInfo",
        "parameters": [
          { "name": "code", "in": "path", "required": true, "schema": { "type": "string" }, "description": "股票代码（6位）", "example": "600519" }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "traceId": { "type": "string" },
                    "code": { "type": "string", "example": "600519" },
                    "name": { "type": "string", "example": "贵州茅台" },
                    "summary": { "type": "string" },
                    "latest_reports": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "title": { "type": "string" },
                          "broker": { "type": "string" },
                          "date": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error400" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Session": {
        "type": "object",
        "properties": {
          "session_id": { "type": "string", "format": "uuid" },
          "title": { "type": "string", "example": "新会话" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" },
          "query_count": { "type": "integer", "example": 0 }
        }
      },
      "QARecord": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "rec_a1b2c3d4e5f6" },
          "session_id": { "type": "string", "format": "uuid" },
          "query": { "type": "string" },
          "answer": { "type": "string" },
          "llm_used": { "type": "boolean" },
          "model": { "type": "string", "nullable": true },
          "response_time_ms": { "type": "integer" },
          "answer_source": { "type": "string", "enum": ["copaw", "bailian", "demo"] },
          "stocks": { "type": "array", "items": { "type": "object" } },
          "timestamp": { "type": "string", "format": "date-time" }
        }
      }
    },
    "responses": {
      "Error400": {
        "description": "请求错误",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "string", "example": "EMPTY_QUERY" },
                    "message": { "type": "string", "example": "请输入问题" },
                    "details": { "type": "object" },
                    "traceId": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}